+++ database/migrations/2016_10_22_134445_recruiters.php
@@ -1,0 +1,41 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class Recruiters extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('recruiters', function (Blueprint $table) {
+            $table->uuid('id')->primary();
+            $table->string('name')->unique();
+            $table->string('url')->nullable();
+            $table->timestamps();
+        });
+
+        Schema::table('searches', function (Blueprint $table) {
+            $table->boolean('no_recruiters')->default(false);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('searches', function (Blueprint $table) {
+            $table->dropColumn('no_recruiters');
+        });
+
+        Schema::dropIfExists('recruiters');
+    }
+}