+++ resources/views/layouts/app.blade.php
@@ -13,6 +13,9 @@
         </div>
 
         @include('layouts.footer')
-
+        <script type="text/javascript" src="/js/jquery.min.js"></script>
+        <script type="text/javascript" src="/js/tether.min.js"></script>
+        <script type="text/javascript" src="/js/bootstrap.min.js"></script>
+        <script type="text/javascript" src="/js/app.js"></script>
     </body>
 </html>